Рефакторинг работы с файлами

+++ src/BulletinBoard/Application/BulletinBoard.Application.AppServices/Contexts/Attachment/Services/AttachmentService.cs
@@ -1,5 +1,5 @@
 ﻿using BulletinBoard.Application.AppServices.Contexts.Attachment.Repositories;
-using BulletinBoard.Application.AppServices.Contexts.User.Repositories;
+using BulletinBoard.Application.AppServices.Contexts.Post.Repositories;
 using BulletinBoard.Contracts.Attachment;
 
 namespace BulletinBoard.Application.AppServices.Contexts.Attachment.Services
@@ -8,16 +8,16 @@ namespace BulletinBoard.Application.AppServices.Contexts.Attachment.Services
     public class AttachmentService : IAttachmentService
     {
         private readonly IAttachmentRepository _attachmentRepository;
-        //private readonly IUserRepository _userRepository;
+        private readonly IPostRepository _postRepository;
 
         /// <summary>
         /// Инициализация экземпляра <see cref="AttachmentService"/>
         /// </summary>
         /// <param name="attachmentRepository">Репозиторий для работы с вложениями.</param>
-        public AttachmentService(IAttachmentRepository attachmentRepository/*, IUserRepository userRepository*/)
+        public AttachmentService(IAttachmentRepository attachmentRepository, IPostRepository postRepository)
         {
             _attachmentRepository = attachmentRepository;
-            //_userRepository = userRepository;
+            _postRepository = postRepository;
         }
 
         /// <inheritdoc/>
@@ -33,21 +33,45 @@ namespace BulletinBoard.Application.AppServices.Contexts.Attachment.Services
         }
 
         /// <inheritdoc/>
-        public Task DeleteAsync(Guid id, CancellationToken cancellationToken)
+        public async Task<bool> DeleteAsync(Guid id, CancellationToken cancellationToken)
         {
-            return _attachmentRepository.DeleteAsync(id, cancellationToken);
+            var existedEntity = await _attachmentRepository.GetInfoByIdAsync(id, cancellationToken);
+            if (existedEntity == null) 
+            {
+                return true;
+            }
+
+            await _attachmentRepository.DeleteAsync(id, cancellationToken);
+            return false;
         }
 
         /// <inheritdoc/>
-        public Task<Guid> UploadAsync(AttachmentDto attachment, Guid postId, CancellationToken cancellationToken)
+        public async Task<Guid> UploadAsync(AttachmentDto attachment, Guid postId, CancellationToken cancellationToken)
         {
-             return _attachmentRepository.UploadAsync(attachment, postId, cancellationToken);
+            var post = await _postRepository.GetByIdAsync(postId, cancellationToken);
+            if (post == null)
+            {
+                return Guid.Empty;
+            }
+
+            if (post.Attachments.Count >= 10)
+            {
+                return Guid.Empty;
+            }
+
+            return await _attachmentRepository.UploadAsync(attachment, postId, cancellationToken);
         }
 
         /// <inheritdoc/>
-        public Task<AttachmentDto?> DownloadAsync(Guid id, CancellationToken cancellationToken)
+        public async Task<AttachmentDto?> DownloadAsync(Guid id, CancellationToken cancellationToken)
         {
-            return _attachmentRepository.DownloadAsync(id, cancellationToken);
+            var existedEntity = await _attachmentRepository.GetInfoByIdAsync(id, cancellationToken);
+            if (existedEntity == null)
+            {
+                return null;
+            }
+
+            return await _attachmentRepository.DownloadAsync(id, cancellationToken);
         }
     }
 }
